Rust: Split key into multiple modules
