+++ src/bindings/rust/elektra/src/key/mod.rs
@@ -1,5 +1,13 @@
-pub mod key;
-pub mod keybuilder;
+mod error;
+mod key;
+mod builder;
+mod readable;
+mod writable;
+mod readonly;
 
-pub use self::key::{BinaryKey, KeyError, ReadableKey, StringKey, WriteableKey};
-pub use self::keybuilder::KeyBuilder;
+pub use self::readonly::ReadOnly;
+pub use self::error::KeyError;
+pub use self::key::{BinaryKey, StringKey};
+pub use self::builder::KeyBuilder;
+pub use self::readable::ReadableKey;
+pub use self::writable::WriteableKey;
