+++ src/bindings/rust/elektra/src/lib.rs
@@ -2,5 +2,6 @@ extern crate elektra_sys;
 
 pub mod key;
 
-pub use self::key::keybuilder::KeyBuilder;
-pub use self::key::{BinaryKey, KeyError, ReadableKey, StringKey, WriteableKey};
+pub use self::key::{
+    BinaryKey, KeyBuilder, KeyError, ReadOnly, ReadableKey, StringKey, WriteableKey,
+};
